(Mtc) Fix XML generation failure when ussd-string is missing
This fix addresses an issue where the XML is not generated when only the error-code element is present without a ussd-string.

During a USSI operation, if the UE rejects the USSD, it must send an INFO message containing USSD data with an error-code. Currently, the XML generation fails in this scenario because the ussd-string is absent.

Bug: 468979445
Test: atest ImsStackNativeEnablerMtcTests
Test: simulation by TISS for USSD cancelation
Flag: NONE pre-production
Change-Id: I880405f8c73b7ee27e9b2705d69ccafb1ee8f218