(Mtc) Notify the Telephony when cross sim status is updated
In order to display the "Shared data call" text on the UI even during a call, that information must be sent to Telephony whenever the cross-SIM status changes.
This information is included in `CallInfo` and passed to the upper layer.
Regardless of SIP signaling, it seems acceptable for `ImsCallSessionListener.callSessionUpdated(ImsCallProfile profile)` to be called whenever the call information changes.

Flag: NONE pre-production
Bug: 465868627
Test: atest ImsStackNativeTests
Test: atest ImsStackJavaTests
Test: simulated with TISS for calling `onCallInfoChanged`
Change-Id: Ie89e90b85753e1eccf4c002b4c1f89365a062b7b