(Mtc) Bug fix - conference subscrition on participant is not triggered
This fix addresses an issue where conference participants do not initiate a subscription.

According to the imsvoice.enable_conference_subscribe_by_participant_bool setting, conference participants are required to subscribe to CEP. However, the current code fails to invoke the necessary function. This change ensures that the function is called when a user is updated as a conference participant.

Bug: 471048882
Test: atest ImsStackNativeEnablerMtcTests
Test: simulation using TISS for
      incoming INVITE with isfocus
      incoming re-INVITE with isfocus
      incoming 200 OK with isfocus
      in-dialog, out dialog.
Flag: NONE pre-production
Change-Id: I0b802f91d2294341d09352148833751a3ba1346e