(xml) Apply escaping to special characters before parsing XML
We observed that the XML body in the incoming INFO request contains
an unescaped special character('&') in its text node, leading to XML
parsing failure during USSI procedure.
To mitigate this network-side malformation, we are adding a pre-parsing
step to escape this special character, allowing the XML body to be
parsed correctly.

The following characters should be escaped in a valid XML document:
 - '&', '"' (double quot), ''' (single quot), '<', '>'

Bug: 470841868
Flag: EXEMPT BUGFIX
Test: atest ImsStackNativeTests
Test: manual (verify IMS registration on target device)
Change-Id: I937b2975656403f88b5df87c23e4a6545b684888