(isil) Remove VT Capability during a Voice call hold and restore it on Resume
When a voice call is held by the remote side, Dialer does not disable the video upgrade UI. As a result, a video upgrade request may be sent from the local side, which can cause unexpected behavior for the call. In addition, if UE receives a video upgrade request while the voice call is held locally, the UE may become stuck in the video upgrade process, as Dialer is unable to handle this scenario correctly.

In the current design for KR SIM, the call type in the local call profile is changed from CALL_TYPE_VIDEO_N_VOICE to CALL_TYPE_VOICE when the call is held by either the remote or local side. When the call is resumed, the call type is restored to CALL_TYPE_VIDEO_N_VOICE.
As a result, when the call is held:
- Incoming video upgrade requests are rejected by Android Telephony
- Dialer disables the video upgrade UI due to the lack of VT capability
The VT capability is recovered once the call is resumed.

This change simply removes the KR SIM check from the hold/resume-related code path and add the corresponding test coverage.

Bug: 460575449

Test: Live network testing PASS
Flag: NONE pre-production

Change-Id: Idbc5736dc7288bf69031c04422fac7b15a5812b3